Test image auto update
There seems to be a mechanism to update the operand images in operator,
when operand's image digest is changed. However, we don't know yet this
feature is working properly or even for which image registry it is triggered.

This commit adds some candidate image registries to see which will be updated.
Once the test is done, this commit has to be reverted by leaving the correct one in here.